requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
lxml==4.9.2
numpy==1.24.1
Paginator==0.5.1
pandas==1.5.3
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
uWSGI==2.0.21
